Remove rsync delta (v2) support
Remove resumable-request, docker-delta dependencies. Docker-delta is only used for v2 rsync deltas
and is not needed for v3. resumable-request is not used elsewhere either. Remove isV2Delta function,
as this PR is designed to be merged with balenaOS 3. v3's supervisor will not allow old v2 deltas,
so at that point, the dependency removals will be safe to merge.

Change-type: minor
Connects-to: #1602
Signed-off-by: Christina Wang <christina@balena.io>